requirements.txt
# Core dependencies
redis>=5.0.0
PyPDF2>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
django>=4.2.0
requests>=2.31.0
openpyxl>=3.1.0

# Database
mysql-connector-python>=8.0.22
mysqlclient>=2.0.2
PyMySQL>=0.10.1

# Task Scheduling and Messaging
APScheduler>=3.8.1
kafka-python>=2.0.2

# Media Processing
moviepy>=1.0.3

# HTTP and API
httpx>=0.24.1
pydantic>=2.0.3

# Utils
protobuf>=3.19.0
six>=1.15.0
lxml>=4.9.3
beautifulsoup4>=4.9.3
urllib3>=1.25.4,<1.27
boto3>=1.24.77
pytz>=2021.3

# Redis and caching
django-redis>=5.4.0
redis-py-cluster>=2.1.3

# File handling and processing
python-magic>=0.4.27
filetype>=1.2.0
chardet>=5.2.0

# Lark SDK and API
larksuiteoapi>=2.0.7
pycryptodome>=3.19.0

# Background tasks and scheduling
celery>=5.3.6
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Testing and development
pytest>=7.4.3
pytest-django>=4.7.0
pytest-asyncio>=0.21.1
fakeredis>=2.20.0

# Production deployment
gunicorn>=21.2.0
whitenoise>=6.6.0
django-cors-headers>=4.3.1

# Monitoring and logging
sentry-sdk>=1.39.1
python-json-logger>=2.0.7 